refactor(condo): DOMA-8810 remove TicketCommentsTime schema
