--- apps/condo/domains/ticket/access/TicketCommentsTime.js
@@ -1,45 +1,0 @@
-/**
- * Generated by `createschema ticket.TicketCommentsTime 'organization:Relationship:Organization:CASCADE; ticket:Relationship:Ticket:CASCADE; lastCommentAt:DateTimeUtc; lastResidentCommentAt:DateTimeUtc;'`
- */
-
-const { throwAuthenticationError } = require('@open-condo/keystone/apolloErrorFormatter')
-
-const { queryOrganizationEmployeeFor, queryOrganizationEmployeeFromRelatedOrganizationFor } = require('@condo/domains/organization/utils/accessSchema')
-const { RESIDENT } = require('@condo/domains/user/constants/common')
-
-async function canReadTicketCommentsTimes ({ authentication: { item: user } }) {
-    if (!user) return throwAuthenticationError()
-    if (user.deletedAt) return false
-
-    if (user.isAdmin || user.isSupport) return {}
-
-    if (user.type === RESIDENT) {
-        return { ticket: { client: { id: user.id }, canReadByResident: true } }
-    }
-
-    return {
-        organization: {
-            OR: [
-                queryOrganizationEmployeeFor(user.id),
-                queryOrganizationEmployeeFromRelatedOrganizationFor(user.id),
-            ],
-        },
-    }
-}
-
-async function canManageTicketCommentsTimes ({ authentication: { item: user } }) {
-    if (!user) return throwAuthenticationError()
-    if (user.deletedAt) return false
-    if (user.isAdmin) return true
-
-    return false
-}
-
-/*
-  Rules are logical functions that used for list access, and may return a boolean (meaning
-  all or no items are available) or a set of filters that limit the available items.
-*/
-module.exports = {
-    canReadTicketCommentsTimes,
-    canManageTicketCommentsTimes,
-}
