+++ apps/condo/domains/ticket/schema/TicketComment.test.js
@@ -37,8 +37,6 @@ const { makeClientWithNewRegisteredAndLoggedInUser, makeClientWithResidentUser }
 
 const { ERRORS: TicketCommmentErrors } = require('./TicketComment')
 
-const { Property } = require('../../property/utils/testSchema')
-
 
 describe('TicketComment', () => {
     let admin
@@ -289,6 +287,56 @@ describe('TicketComment', () => {
                     id: ticket.id,
                 })
 
+                expect(readTicket1.lastResidentCommentAt).toEqual(ticketComment.createdAt)
+                expect(readTicket1.lastCommentAt).toEqual(ticketComment1.createdAt)
+            })
+
+            it('update lastCommentWithResidentTypeAt and lastCommentAt of related ticket after create TicketComment with resident type', async () => {
+                const adminClient = await makeLoggedInAdminClient()
+                const [organization] = await createTestOrganization(adminClient)
+                const [property] = await createTestProperty(adminClient, organization)
+
+                const residentClient = await makeClientWithResidentUser()
+                const unitName = faker.random.alphaNumeric(5)
+                const unitType = FLAT_UNIT_TYPE
+                await createTestResident(adminClient, residentClient.user, property, {
+                    unitName,
+                    unitType,
+                })
+
+                const employeeClient = await makeClientWithNewRegisteredAndLoggedInUser()
+                const [role] = await createTestOrganizationEmployeeRole(adminClient, organization, {
+                    canManageTickets: true,
+                    canManageTicketComments: true,
+                })
+                await createTestOrganizationEmployee(adminClient, organization, employeeClient.user, role)
+
+                const [ticket] = await createTestTicket(residentClient, organization, property, {
+                    unitName,
+                    unitType,
+                })
+
+                const [ticketComment] = await createTestTicketComment(residentClient, ticket, residentClient.user, {
+                    type: RESIDENT_COMMENT_TYPE,
+                })
+
+                const readTicket = await Ticket.getOne(residentClient, {
+                    id: ticket.id,
+                })
+
+                expect(readTicket.lastCommentWithResidentTypeAt).toEqual(ticketComment.createdAt)
+                expect(readTicket.lastResidentCommentAt).toEqual(ticketComment.createdAt)
+                expect(readTicket.lastCommentAt).toEqual(ticketComment.createdAt)
+
+                const [ticketComment1] = await createTestTicketComment(employeeClient, ticket, employeeClient.user, {
+                    type: RESIDENT_COMMENT_TYPE,
+                })
+
+                const readTicket1 = await Ticket.getOne(residentClient, {
+                    id: ticket.id,
+                })
+
+                expect(readTicket1.lastCommentWithResidentTypeAt).toEqual(ticketComment1.createdAt)
                 expect(readTicket1.lastResidentCommentAt).toEqual(ticketComment.createdAt)
                 expect(readTicket1.lastCommentAt).toEqual(ticketComment1.createdAt)
             })
@@ -558,14 +606,18 @@ describe('TicketComment', () => {
                 const admin = await makeLoggedInAdminClient()
                 const residentClient = await makeClientWithResidentUser()
 
+                const [organization] = await createTestOrganization(admin)
+                const [property] = await createTestProperty(admin, organization)
+                const unitName = faker.random.alphaNumeric(5)
                 const content = faker.lorem.sentence()
 
-                const property = await Property.getOne(admin, { id: '254b852b-e6cd-4b7c-844c-c7c1536fe22e' })
-
                 await createTestResident(admin, residentClient.user, property, {
-                    unitName: '1',
+                    unitName,
                 })
-                const [ticketComment] = await createTestTicketComment(residentClient, { id: '0d9bb499-a0fc-4297-a935-d210ff4374b9' }, residentClient.user, {
+                const [ticket] = await createTestTicket(residentClient, organization, property, {
+                    unitName,
+                })
+                const [ticketComment] = await createTestTicketComment(residentClient, ticket, residentClient.user, {
                     type: RESIDENT_COMMENT_TYPE,
                     content,
                 })
