--- apps/condo/domains/ticket/schema/TicketCommentsTime.js
@@ -1,58 +1,0 @@
-/**
- * Generated by `createschema ticket.TicketCommentsTime 'organization:Relationship:Organization:CASCADE; ticket:Relationship:Ticket:CASCADE; lastCommentAt:DateTimeUtc; lastResidentCommentAt:DateTimeUtc;'`
- */
-
-const { Relationship, DateTimeUtc } = require('@keystonejs/fields')
-
-const { historical, versioned, uuided, tracked, softDeleted, dvAndSender } = require('@open-condo/keystone/plugins')
-const { GQLListSchema } = require('@open-condo/keystone/schema')
-
-const { addOrganizationFieldPlugin } = require('@condo/domains/organization/schema/plugins/addOrganizationFieldPlugin')
-const access = require('@condo/domains/ticket/access/TicketCommentsTime')
-
-const TicketCommentsTime = new GQLListSchema('TicketCommentsTime', {
-    schemaDoc: 'The time of the last comment and the last comment of the resident in a specific ticket',
-    fields: {
-        ticket: {
-            schemaDoc: 'Link to the ticket',
-            type: Relationship,
-            ref: 'Ticket',
-            isRequired: true,
-            knexOptions: { isNotNullable: true },
-            kmigratorOptions: { null: false, on_delete: 'models.CASCADE' },
-        },
-
-        // Время последнего коммента во вкладке с жителем
-        lastCommentAt: {
-            schemaDoc: 'Time of last comment with resident type in this ticket',
-            type: DateTimeUtc,
-            isRequired: true,
-        },
-
-        // Время последнего коммента жителя во вкладке с жителем
-        lastResidentCommentAt: {
-            schemaDoc: 'Time of last comment with resident user in this ticket',
-            type: DateTimeUtc,
-        },
-    },
-    plugins: [
-        addOrganizationFieldPlugin({ fromField: 'ticket', isRequired: true }),
-        uuided(),
-        versioned(),
-        tracked(),
-        softDeleted(),
-        dvAndSender(),
-        historical(),
-    ],
-    access: {
-        read: access.canReadTicketCommentsTimes,
-        create: access.canManageTicketCommentsTimes,
-        update: access.canManageTicketCommentsTimes,
-        delete: false,
-        auth: true,
-    },
-})
-
-module.exports = {
-    TicketCommentsTime,
-}
