+++ apps/condo/domains/ticket/utils/clientSchema/index.ts
@@ -16,7 +16,6 @@ export * as TicketClassifier from './TicketClassifier'
 export * as TicketComment from './TicketComment'
 export * as TicketFilterTemplate from './TicketFilterTemplate'
 export * as TicketCommentFile from './TicketCommentFile'
-export * as TicketCommentsTime from './TicketCommentsTime'
 export * as UserTicketCommentReadTime from './UserTicketCommentReadTime'
 export * as TicketPropertyHint from './TicketPropertyHint'
 export * as TicketPropertyHintProperty from './TicketPropertyHintProperty'
